Fix(client): release the task when data streaming ends
This is also a very important fix, by watching the tokio console with wrk, no more tasks leaked.

```
wrk -c 100 -d10s http://localhost:6611 -H 'Host: foo.com'
```
