+++ tests/e2e/test_http_tunnel_with_subdomain.sh
@@ -5,8 +5,6 @@ root_dir=$(git rev-parse --show-toplevel)
 cur_dir=$root_dir/tests/e2e
 source $cur_dir/util.sh
 
-cargo build
-
 # Function to clean up processes
 cleanup() {
   echo "Cleaning up..."
