+++ tunneld-client/Cargo.toml
@@ -9,15 +9,12 @@ tunneld-pkg = { path = "../tunneld-pkg" }
 clap = { version = "4.5.7", features = ["derive"] }
 tokio = { version = "1.10.1", features = ["full"] }
 tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-console-subscriber = "0.3.0"
 anyhow = "1.0.86"
 tonic = "0.11.0"
 tokio-stream = "0.1.15"
 tokio-util = "0.7.11"
 futures = "0.3.30"
 bytes = "1.6.0"
-reqwest = { version = "0.12" }
 
 [[bin]]
 name = "tunnel"
