+++ tunneld-pkg/Cargo.toml
@@ -8,6 +8,7 @@ tunneld-protocol = { path = "../tunneld-protocol" }
 tokio = { version = "1.10.1", features = ["full"] }
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
+console-subscriber = "0.3.0"
 anyhow = "1.0.86"
 tonic = "0.11.0"
 tokio-stream = "0.1.15"
