+++ tunneld-pkg/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod event;
 pub mod io;
+pub mod otel;
 pub mod shutdown;
 pub mod util;
