+++ tunneld-server/Cargo.toml
@@ -13,8 +13,6 @@ tunneld-pkg = { path = "../tunneld-pkg" }
 clap = { version = "4.5.7", features = ["derive"] }
 tokio = { version = "1.10.1", features = ["full"] }
 tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-console-subscriber = "0.3.0"
 async-trait = "0.1.80"
 anyhow = "1.0.86"
 tonic = "0.11.0"
